WIP: Implement execution loop and direct 64-bit pointer support
- Added execution loop in ExecutionBootstrap to run interpreter iteratively
- Implemented 64-bit EIP storage for direct memory mode on 64-bit hosts
- Modified X86_32GuestContext to store full 64-bit entry points via SetEIP64()
- Updated OptimizedX86Executor to use stored 64-bit addresses
- Simplified write() syscall for direct memory mode
- Improved address space handling in direct memory mode

Status: Execution loop starts but hangs on first instruction execution.
Need to debug OptimizedX86Executor or use different interpreter approach.

TODO:
- Debug LoadInstr() or use alternative instruction reading mechanism
- Consider simpler interpreter for initial testing

Amp-Thread-ID: https://ampcode.com/threads/T-019c2824-dde0-707e-a060-a22818d247ae

+++ ExecutionBootstrap.h
@@ -20,7 +20,7 @@ private:
 		ElfImage *image;
 		void *stackBase;
 		size_t stackSize;
-		uint32 entryPoint;
+		uintptr_t entryPoint;  // Use uintptr_t to store full 64-bit pointers on 64-bit hosts
 		uint32 stackPointer;
 		DynamicLinker *linker;  // For loading dependencies
 	};
